add divisions search by name part (LIKE)
